fixing rts

+++ public/js/dashboard.js
@@ -3,24 +3,30 @@
 async function createBlog(event) {
   event.preventDefault();
 
-  const name = document.querySelector('input[name="blog-name"]').value.trim();
-  const description = document.querySelector('input[name="blog-description"]').value.trim();
+  let inputName = document.querySelector('#blog-name')
+  let inputDesc = document.querySelector('#blog-description')
 
-  const response = await fetch(`/api/blogs`, {
-    method: 'POST',
-    body: JSON.stringify({
-      name,
-      description
-    }),
-    headers: {
-      'Content-Type': 'application/json'
+  const name = inputName.value.trim();
+  const description = inputDesc.value.trim();
+
+  if (name && description) {
+
+    const response = await fetch(`/api/blog`, {
+      method: 'POST',
+      body: JSON.stringify({
+        name,
+        description
+      }),
+      headers: {
+        'Content-Type': 'application/json'
+      }
+    });
+
+    if (response.ok) {
+      document.location.replace('/dashboard');
+    } else {
+      alert(response.statusText);
     }
-  });
-
-  if (response.ok) {
-    document.location.replace('/dashboard');
-  } else {
-    alert(response.statusText);
   }
 }
 
@@ -32,28 +38,31 @@ document.querySelector('.new-blog-form').addEventListener('submit', createBlog);
 async function editBlog(event) {
   event.preventDefault();
 
-  const name = document.querySelector('input[name="blog-name"]').value;
-  const description = document.querySelector('input[name="blog-description"]').value;
-  const id = window.location.toString().split('/')[
-      window.location.toString().split('/').length - 1
-    ];
+  let inputName = document.querySelector('#blog-name')
+  let inputDesc = document.querySelector('#blog-description')
 
-  const response = await fetch(`/api/blogs/${id}`, {
-      method: 'PUT',
-      body: JSON.stringify({
-          name,
-          description
-      }),
-      headers: {
-          'Content-Type': 'application/json'
-      }
-    });
-    
-    if (response.ok) {
-      document.location.replace('/dashboard/');
-    } else {
-      alert(response.statusText);
+  const name = inputName.value;
+  const description = inputDesc.value;
+  // const id = window.location.toString().split('/')[
+  //     window.location.toString().split('/').length - 1
+  //   ];
+
+  const response = await fetch(`/api/blog/${id}`, {
+    method: 'PUT',
+    body: JSON.stringify({
+      name,
+      description
+    }),
+    headers: {
+      'Content-Type': 'application/json'
     }
+  });
+
+  if (response.ok) {
+    document.location.replace('/dashboard/');
+  } else {
+    alert(response.statusText);
+  }
 }
 
 document.querySelector('#update-blog-btn').addEventListener('click', editBlog);
@@ -63,27 +72,23 @@ document.querySelector('#update-blog-btn').addEventListener('click', editBlog);
 
 async function deleteBlog(event) {
   event.preventDefault();
-  
-  const id = window.location.toString().split('/')[
-      window.location.toString().split('/').length - 1
-    ];
 
-  const response = await fetch(`/api/blogs/${id}`, {
+  if (event.target.hasAttribute('data-id')) {
+    const id = event.target.getAttribute('data-id');
+
+    const response = await fetch(`/api/blog/${id}`, {
       method: 'DELETE',
-      body: JSON.stringify({
-        blog_id: id
-      }),
-      headers: {
-        'Content-Type': 'application/json'
-      }
     });
-    
+
     if (response.ok) {
       document.location.replace('/dashboard');
     } else {
-      alert(response.statusText);
+      alert('Failed to delete blog');
     }
-  
+  }
+
+
+
 }
 
 document.querySelector('#delete-blog-btn').addEventListener('click', deleteBlog);
